requirements.txt
streamlit
fpdf2
PyPDF2
python-dotenv